requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alabaster==1.0.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
async-generator==1.10
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bibtexparser==1.4.4
blinker==1.9.0
blis==1.3.3
cachetools==6.2.6
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.4
cymem==2.0.13
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
docutils==0.22.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fake-useragent==2.2.0
filelock==3.20.3
fonttools==4.61.1
fpdf2==2.8.5
free_proxy==1.1.3
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
hyperframe==6.1.0
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
postgrest==2.27.2
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pyroaring==1.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.1
roman-numerals==4.1.0
rpds-py==0.30.0
safetensors==0.7.0
scholarly==1.7.11
scikit-learn==1.8.0
scipy==1.17.0
selenium==4.40.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.5.2
storage3==2.27.2
streamlit==1.53.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
transformers==5.0.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.0
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
